refactor(nav): extract SituationSensor from ExperimentalControls
Third decomposition step (moved before the runner per review — it owns the
legit-pose/context-snapshot state the runner settle writes). Moves the per-tick
situation eval into a SituationSensor service: the legit-pose snapshot, the
recovery cue (show/hide hysteresis), and the context snapshot, all derived from
one idle-gated enclosure ray — plus reseedLegitPose / refreshContextSnapshot /
enclosureProbe(At) and the _lastLegitPose / _contextSnapshot / cue / idle-gate
state.

The idle gate reads engine state via a new ctx.isCameraBusy() predicate; the
recovery-cue dispatch fires on the controls instance via a curated ctx.dispatch
bound callback (not the removed controls god-handle). The break-it idle-gate
proxy patches _updateLegitSnapshotAndCue, kept as an O delegator to
this._sensor.update(), so no harness change was needed and it still reds.

ctx grew: center/latch/streetLevelEnabled/sensor getters + dispatch +
isCameraBusy. Behaviour-preserving: 302/302 nav tests green; lint clean.
